#820 rename the file EventFormFieldsExpaned.tsx to EventFormFieldsExpanded.tsx
