+++ src/components/Event/EventFormFields.tsx
@@ -8,7 +8,7 @@ import { useScreenSizeDetection } from '@/useScreenSizeDetection'
 import { CalendarSelectField } from './fields/CalendarSelectField'
 import { EventDateTimeField } from './fields/EventDateTimeField'
 import { VideoConferenceField } from './fields/VideoConferenceField'
-import { EventFormFieldsExpanded } from './components/EventFormFieldsExpaned'
+import { EventFormFieldsExpanded } from './components/EventFormFieldsExpanded'
 import { useAutoFocusTitle } from './hooks/useAutoFocusTitle'
 import LocationField from './fields/LocationField'
 import { EventFormFieldsProps } from './EventFormFields.types'
